pythonback/requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2-binary==2.9.9
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0
